Remove unnecessary (int) casts before alloc() calls
alloc() already accepts size_t, so (int) casts on size_t values are
redundant and could theoretically cause truncation on values > INT_MAX.

Remove the casts and change alloc_cmdbuff() signature from int to
size_t to match.

Note: list_alloc_with_items() keeps its int parameter since lv_len and
lv_with_items are int, and the call site already has an INT_MAX guard.

Closes #19888

+++ src/memline.c
@@ -3674,7 +3674,7 @@ ml_replace_len(
 	    size_t textproplen = curbuf->b_ml.ml_line_len - oldtextlen;
 
 	    // Need to copy over text properties, stored after the text.
-	    newline = alloc(len + (int)textproplen);
+	    newline = alloc(len + textproplen);
 	    if (newline != NULL)
 	    {
 		mch_memmove(newline, line, len);
